+++ public-api/src/main/kotlin/io/prizy/publicapi/port/contest/mapper/CreateMerchantDtoMapper.kt
@@ -14,7 +14,7 @@ object CreateMerchantDtoMapper {
     return Merchant.builder()
       .name(dto.name)
       .siren(dto.siren)
-      .capital(dto.capital)
+      .capital(dto.capital.toFloat())
       .address(dto.address)
       .logoAssetId(dto.logoAssetId)
       .build()
